Updated the testing scripts
aproxTest.sh = tests pmt for approximate matching, calling it once for each pattern in a file, with error = pat_size - 1.
singlePatternTest.sh = tests pmt for exact matching, calling it once for each patter in a file.
For both scripts above the user must enter the desired algorithm after calling the script, using the  expected format for pmt. The text file is defined in the code.

singleGrepTest.sh = tests grep using a pattern file and a text file, which are defined in the code.
filePatternTest.sh = tests pmt using the -p patfile option, where the text file name and patfile are declared in the code.